+++ src/Accounts/Accounts/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Disable context auto saving when AzureRmContext.json not available
+* Update the reference to Azure Powershell Common to 1.3.5-preview
 
 ## Version 1.7.0
 * Updated Add-AzEnvironment and Set-AzEnvironment to accept parameters AzureAttestationServiceEndpointResourceId and AzureAttestationServiceEndpointSuffix
@@ -33,7 +35,6 @@
 
 ## Version 1.6.4
 * Add a deprecation message for `Resolve-Error` alias.
-* Disable context auto saving when AzureRmContext.json not available
 
 ## Version 1.6.3
 * Update telemetry and url rewriting for generated modules, fix windows unit tests.